[ISSUE #20] try to pass CI
