+++ build.rs
@@ -1,0 +1,4 @@
+fn main() {
+    //link hook dylib
+    println!("cargo:rustc-link-lib=dylib=hook");
+}
