--- open-coroutine/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "open-coroutine"
-version = "0.0.4"
-edition = "2021"
-authors = ["zhangzicheng@apache.org"]
-description = "The open-coroutine is a simple, efficient and generic stackful-coroutine library."
-repository = "https://github.com/acl-dev/open-coroutine/tree/dev/open-coroutine"
-license = "LGPL-3.0 OR Apache-2.0"
-readme = "README.md"
-keywords = ["coroutine", "fiber", "stackful", "hook"]
-
-[dependencies]
-base-coroutine = { version = "0.0.1", path = "../base-coroutine" }
